backend/auth_service/requirements.txt
pydantic[email]
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pymongo==4.6.1
motor==3.3.2 
redis==5.0.1
python-dotenv==1.0.0
python-jose==3.3.0 
cryptography==41.0.7
passlib
python-logstash
python-json-logger
